Move visualize detections to utils
